Consolidate function to add parquet options to state.
